pathway-service/requirements.txt
pathway-ai
pandas
